+++ synch/reader/__init__.py
@@ -1,10 +1,12 @@
 import abc
 import logging
 import signal
+import time
 from signal import Signals
 from typing import Callable, Tuple, Union
 
 from synch.broker import Broker
+from synch.common import insert_log
 from synch.settings import Settings
 
 logger = logging.getLogger("synch.reader")
@@ -13,8 +15,11 @@ logger = logging.getLogger("synch.reader")
 class Reader:
     cursor = None
     fix_column_type = False
+    last_time = 0
+    count = {}
 
     def __init__(self, alias: str):
+        self.alias = alias
         source_db = Settings.get_source_db(alias)
         self.source_db = source_db
         self.host = source_db.get("host")
@@ -35,7 +40,7 @@ class Reader:
         raise NotImplementedError
 
     @abc.abstractmethod
-    def start_sync(self, broker: Broker, insert_interval: int):
+    def start_sync(self, broker: Broker):
         raise NotImplementedError
 
     @abc.abstractmethod
@@ -45,3 +50,18 @@ class Reader:
     @abc.abstractmethod
     def get_source_select_sql(self, schema: str, table: str, sign_column: str = None):
         raise NotImplementedError
+
+    def after_send(self, schema, table):
+        now = int(time.time())
+        schema_table = f"{schema}.{table}"
+        self.count.setdefault(schema_table, 0)
+        self.count[schema_table] += 1
+        if self.last_time == 0:
+            self.last_time = now
+        if now - self.last_time >= Settings.insert_interval():
+            for schema_table, num in self.count.items():
+                logger.info(f"success send {num} events for {schema_table}")
+                s, t = schema_table.split(".")
+                insert_log(self.alias, s, t, num, 1)
+            self.last_time = 0
+            self.count = {}