+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.6
 
+### 0.6.6
+
+- Add `monitoring` in config, which insert monitoring records in ClickHouse.
+
 ### 0.6.5
 
 - Fix bugs in skip database table.